Fixed Most warnings.

+++  saving-face --username 1900Geek@gmail.com/src/sf_model.cpp
@@ -191,21 +191,21 @@ namespace SF
 		person_info.suffix = strtok_s(NULL, "&",&next_token);
 		person_info.emailAddress = strtok_s(NULL, "&",&next_token);
 		
-		model_info.deltaX = atof(strtok_s(NULL,"&",&next_token));
-		model_info.xMax = atof(strtok_s(NULL,"&",&next_token));
-		model_info.xMin = atof(strtok_s(NULL,"&",&next_token));
+		model_info.deltaX = (float)atof(strtok_s(NULL,"&",&next_token));
+		model_info.xMax = (float)atof(strtok_s(NULL,"&",&next_token));
+		model_info.xMin = (float)atof(strtok_s(NULL,"&",&next_token));
 		model_info.xOffset = atoi(strtok_s(NULL,"&",&next_token));
 		model_info.xWidth = atoi(strtok_s(NULL,"&",&next_token));
 
-		model_info.deltaY = atof(strtok_s(NULL,"&",&next_token));
-		model_info.yMax = atof(strtok_s(NULL,"&",&next_token));
-		model_info.yMin = atof(strtok_s(NULL,"&",&next_token));
+		model_info.deltaY = (float)atof(strtok_s(NULL,"&",&next_token));
+		model_info.yMax = (float)atof(strtok_s(NULL,"&",&next_token));
+		model_info.yMin = (float)atof(strtok_s(NULL,"&",&next_token));
 		model_info.yOffset = atoi(strtok_s(NULL,"&",&next_token));
 		model_info.yWidth = atoi(strtok_s(NULL,"&",&next_token));
 
-		model_info.deltaZ = atof(strtok_s(NULL,"&",&next_token));
-		model_info.zMax = atof(strtok_s(NULL,"&",&next_token));
-		model_info.zMin = atof(strtok_s(NULL,"&",&next_token));
+		model_info.deltaZ = (float)atof(strtok_s(NULL,"&",&next_token));
+		model_info.zMax = (float)atof(strtok_s(NULL,"&",&next_token));
+		model_info.zMin = (float)atof(strtok_s(NULL,"&",&next_token));
 		model_info.zWidth = atoi(strtok_s(NULL,"&",&next_token));
 		model_info.muid = atol(strtok_s(NULL,"&",&next_token));
 		char temp[2];
